requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
email-validator==2.1.1
SQLAlchemy==2.0.49
alembic==1.13.1
psycopg2-binary==2.9.11
python-jose==3.3.0
python-multipart==0.0.26
